Fix brudervn reader. Added power measurements to volume benchmarks.

+++ trrojand3d11/src/volume_benchmark_base.cpp
@@ -127,14 +127,14 @@ void trrojan::d3d11::volume_benchmark_base::load_brudervn_xfer_func(
         const char *path, ID3D11Device *device, ID3D11Texture1D **outTexture,
         ID3D11ShaderResourceView **outSrv) {
     std::vector<std::uint8_t> data;
-    std::uint8_t value;
+    std::uint32_t value;
 
     std::ifstream file(path, std::ios::in);
     if (file.is_open()) {
         data.reserve(256);
 
         while (file >> value) {
-            data.push_back(value);
+            data.push_back(static_cast<uint8_t>(value));
         }
 
     } else {
